Extract offscreen render utilities
